MINGW64 fixes - 64-bit model is different between Linux (LP64) and Windows (LLP64) only in the size of long type; let's avoid those casts to long as much as possible

+++ trunk/src/gridflow.cxx
@@ -166,7 +166,7 @@ ostream &operator << (ostream &self, const t_atom &a) {
 		case A_FLOAT:   self << a.a_float; break;
 		case A_SYMBOL:  self << a.a_symbol->s_name; break; // i would rather show backslashes here...
 		case A_DOLLSYM: self << a.a_symbol->s_name; break; // for real, it's the same thing as A_SYMBOL in pd >= 0.40
-		case A_POINTER: self << "\\p(0x" << std::hex << (unsigned long)a.a_gpointer << std::dec << ")"; break;
+		case A_POINTER: self << "\\p(0x" << std::hex << (uintptr_t)a.a_gpointer << std::dec << ")"; break;
 		case A_COMMA:   self << ","; break;
 		case A_SEMI:    self << ";"; break;
 		case A_DOLLAR:  self << "$" << a.a_w.w_index; break;
